update controller & response

+++ src/main/java/com/example/traffic_signal/services/TrafficSignalService.java
@@ -1,11 +1,11 @@
 package com.example.traffic_signal.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.example.traffic_signal.TrafficSignalApplication;
 import com.example.traffic_signal.module.TrafficSignal;
 import com.example.traffic_signal.repository.TrafficSignalrepo;
 
@@ -19,21 +19,33 @@ public class TrafficSignalService  {
 		return trafficSignalrepo.findAll();
 	}
 	
-	public TrafficSignal createSignal(TrafficSignal newSignal) {
-		return trafficSignalrepo.save(newSignal);
+	public String createSignal(TrafficSignal trafficSignal) {
+		trafficSignalrepo.save(trafficSignal);
+		return "signal created";
 	}
 	
-	public TrafficSignal getSignalsById(int signalId) {
-		return trafficSignalrepo.findById(signalId).get();
+	public TrafficSignal  getSignalsById(long signalId) {
+		TrafficSignal  signal= trafficSignalrepo.findById(signalId).get();
+		return signal;
 	}
 	
-	
-	public TrafficSignal updateSignal(int signalId,TrafficSignal traffic_Signal) {
-		return trafficSignalrepo.save(traffic_Signal);
+	public boolean updateSignal(long signalId,TrafficSignal signal) {
+		Optional<TrafficSignal> signals=trafficSignalrepo.findById(signalId);
+		if(signals.isPresent() && signal.getSignalId()==signalId) {
+			trafficSignalrepo.save(signal);
+			return true;
+		}
+		return false;
+		
 	}
 	
-	public void deleteSignalById(int signalId) {
-		trafficSignalrepo.deleteById(signalId);
+	public boolean deleteSignalById(long signalId) {
+		Optional< TrafficSignal> signal=trafficSignalrepo.findById(signalId);
+		if(signal.isPresent()) {
+			 trafficSignalrepo.deleteById(signalId);
+			return true; 
+		}
+		return false;
 	}
 	public void deleteSignals() {
 		trafficSignalrepo.deleteAll();
